requirements.txt
absl-py
numpy
jax
jaxtyping
einops
flax
orbax
chex
setuptools
plotext
rich >= 12.4

#The tests
tensorflow